backport: drop go-gh/v2 on backport branches via magefile split
Move ReportFailedTests and RequiresUpdate (the two mage targets that
import go-gh/v2 via dev/testsreporter and dev/requiresupdate) into a
new magefile_main_only.go. These targets are only invoked from
main-branch pipelines; backport branches never call them.

Update backport_branch.sh to exclude this file and the two dev packages
when creating a backport branch, so go mod tidy drops go-gh/v2. This
allows backport branches to retain their original Go version — .go-version
is no longer copied from main.

go mod tidy now runs in a subshell that installs Go from the backport
branch's own .go-version via gvm; main's Go version is restored in the
parent shell afterwards.
